chore(core): description module
Split off a new module for text descriptions.
